Query the dataset id with an http request
There are some problems generating the same UUID from name in Python
(Flocker) and in Go. For that reason a new query to lookup for the
dataset id given the metadata-name was added on this commit.

If the problem generating the same UUIDs in Flocker CLI than in Go,
probably this commit can be reverted.